+++ core-addon/Core.js
@@ -4,7 +4,11 @@
 
 import Storage from "./Storage.js";
 import DataCollection from "./DataCollection.js";
+import Glean from "@mozilla/glean/webext";
 import * as rallyMetrics from "../public/generated/rally.js";
+import * as enrollmentMetrics from "../public/generated/enrollment.js";
+import * as unenrollmentMetrics from "../public/generated/unenrollment.js";
+import * as rallyPings from "../public/generated/pings.js";
 
 // The path of the embedded resource used to control options.
 const OPTIONS_PAGE_PATH = "public/index.html";
@@ -285,6 +289,12 @@ export default class Core {
       throw new Error(`Core._handleExternalMessage - unexpected sender ${sender.id}`);
     }
 
+    const knownStudy = availableStudies.find(s => s.addonId == sender.id);
+    if (!("schemaNamespace" in knownStudy)) {
+      return Promise.reject(
+        new Error(`Core._handleExternalMessage - No schema namespace specified in remote settings for ${sender.id}`));
+    }
+
     let joinedStudies = await this._storage.getActivatedStudies();
     if (!joinedStudies.includes(sender.id)) {
       throw new Error(`Core._handleExternalMessage - ${sender.id} not joined`);
@@ -308,10 +318,11 @@ export default class Core {
         if (pausedStudies.includes(sender.id)) {
           throw new Error(`Core._handleExternalMessage - ${sender.id} is paused and may not send data`);
         }
-        const { payloadType, payload, namespace, keyId, key } = message.data;
+
+        const { payloadType, payload, keyId, key } = message.data;
         let rallyId = await this._storage.getRallyID();
         return await this._dataCollection.sendPing(
-          rallyId, payloadType, payload, namespace, keyId, key
+          rallyId, payloadType, payload, knownStudy.schemaNamespace, keyId, key
         );
       }
       default:
@@ -408,7 +419,10 @@ export default class Core {
     // Override the uninstall URL to include the rallyID, for deleting data without exposing the Rally ID.
     await this.setUninstallURL();
 
-    // Finally send the ping.
+    rallyPings.enrollment.submit();
+
+    // Finally send the ping. Important: remove this line once the migration
+    // to Glean.js is finally complete.
     await this._dataCollection.sendEnrollmentPing(rallyId, undefined, deletionId);
   }
 
@@ -427,13 +441,22 @@ export default class Core {
       return Promise.reject(
         new Error(`Core._enrollStudy - Unknown study ${studyAddonId}`));
     }
+    const knownStudy = knownStudies.find(s => s.addonId == studyAddonId);
+    if (!("schemaNamespace" in knownStudy)) {
+      return Promise.reject(
+        new Error(`Core._enrollStudy - No schema namespace specified in remote settings for ${studyAddonId}`));
+    }
 
     // Record that user activated this study.
     await this._storage.appendActivatedStudy(studyAddonId);
 
-    // Finally send the ping.
+    enrollmentMetrics.studyId.set(studyAddonId);
+    rallyPings.studyEnrollment.submit();
+
+    // Finally send the ping. Important: remove this line once the migration
+    // to Glean.js is finally complete.
     let rallyId = await this._storage.getRallyID();
-    await this._dataCollection.sendEnrollmentPing(rallyId, studyAddonId);
+    await this._dataCollection.sendEnrollmentPing(rallyId, knownStudy.schemaNamespace);
   }
 
   /**
@@ -454,6 +477,12 @@ export default class Core {
         new Error(`Core._unenrollStudy - Unknown study ${studyAddonId}`));
     }
 
+    const knownStudy = knownStudies.find(s => s.addonId == studyAddonId);
+    if (!("schemaNamespace" in knownStudy)) {
+      return Promise.reject(
+        new Error(`Core._enrollStudy - No schema namespace specified in remote settings for ${studyAddonId}`));
+    }
+
     // Attempt to send an uninstall message, but move on if the
     // delivery fails: studies will not be able to send anything
     // without the Core Add-on anyway. Moreover, they might have been
@@ -466,8 +495,13 @@ export default class Core {
 
     await this._storage.removeActivatedStudy(studyAddonId);
 
+    unenrollmentMetrics.studyId.set(studyAddonId);
+    rallyPings.studyUnenrollment.submit();
+
+    // Important: remove these lines once the migration
+    // to Glean.js is finally complete.
     let rallyId = await this._storage.getRallyID();
-    await this._dataCollection.sendDeletionPing(rallyId, studyAddonId);
+    await this._dataCollection.sendDeletionPing(rallyId, knownStudy.schemaNamespace);
   }
 
   /**
@@ -481,7 +515,8 @@ export default class Core {
    */
   async _unenroll() {
     // Uninstall all known studies that are still installed.
-    let installedStudies = (await this._availableStudies)
+    const availableStudies = await this._availableStudies;
+    let installedStudies = availableStudies
       .filter(s => s.studyInstalled)
       .map(s => s.addonId);
     for (let studyId of installedStudies) {
@@ -502,7 +537,16 @@ export default class Core {
     // for each of them.
     let studyList = await this._storage.getActivatedStudies();
     for (let studyId of studyList) {
-      await this._dataCollection.sendDeletionPing(rallyId, studyId);
+      unenrollmentMetrics.studyId.set(studyId);
+      rallyPings.studyUnenrollment.submit();
+
+      // Important: remove these lines once the migration
+      // to Glean.js is finally complete.
+      const knownStudy = availableStudies.find(s => s.addonId == studyId);
+      if (!("schemaNamespace" in knownStudy)) {
+        console.error(`Core._handleExternalMessage - No schema namespace specified in remote settings for ${studyId}`);
+      }
+      await this._dataCollection.sendDeletionPing(rallyId, knownStudy.schemaNamespace);
     }
 
     // Clear locally stored IDs.
@@ -511,6 +555,10 @@ export default class Core {
 
     // Clear the list of studies user took part in.
     await this._storage.clearActivatedStudies();
+
+    // Flip upload enabled to disabled: this will trigger a
+    // deletion-request.
+    Glean.setUploadEnabled(false);
 
     // Finally, uninstall the addon.
     await browser.management.uninstallSelf({ showConfirmDialog: false });
@@ -547,7 +595,7 @@ export default class Core {
     // Only do this for "uninstall" messages.
     let studyList = await this._storage.getActivatedStudies();
     if (!studyList.includes(studyId)
-        && type != "uninstall") {
+      && type != "uninstall") {
       return Promise.reject(
         new Error(`Core._sendMessageToStudy - "${studyId}" is not a known study`));
     }