+++ tests/core-addon/unit/Core.test.js
@@ -10,6 +10,8 @@ import Glean from "@mozilla/glean/webext";
 
 import Core from '../../../core-addon/Core.js';
 import * as rallyMetrics from "../../../public/generated/rally.js";
+import * as enrollmentMetrics from "../../../public/generated/enrollment.js";
+import * as rallyPings from "../../../public/generated/pings.js";
 
 
 // The website to post deletion IDs to.
@@ -18,13 +20,17 @@ const OFFBOARD_URL = "https://production.rally.mozilla.org/offboard";
 // A fake study id to use in the tests when looking for a
 // "known" study.
 const FAKE_STUDY_ID = "test@ion-studies.com";
+const FAKE_STUDY_NAMESPACE = "fake-study-installed-namespace"
 const FAKE_STUDY_ID_NOT_INSTALLED = "test-not-installed@ion-studies.com";
+const FAKE_STUDY_NOT_INSTALLED_NAMESPACE = "fake-study-not-installed-namespace"
 const FAKE_STUDY_LIST = [
   {
-    "addonId": FAKE_STUDY_ID
+    "addonId": FAKE_STUDY_ID,
+    "schemaNamespace": FAKE_STUDY_NAMESPACE
   },
   {
-    "addonId": FAKE_STUDY_ID_NOT_INSTALLED
+    "addonId": FAKE_STUDY_ID_NOT_INSTALLED,
+    "schemaNamespace": FAKE_STUDY_NOT_INSTALLED_NAMESPACE
   }
 ];
 const FAKE_UUID = "c0ffeec0-ffee-c0ff-eec0-ffeec0ffeec0";
@@ -206,17 +212,19 @@ describe('Core', function () {
 
       sinon.spy(this.core._dataCollection, "sendEnrollmentPing");
       sinon.spy(this.core._storage, "setRallyID");
+      const enrollmentPingSpy = sinon.spy(rallyPings.enrollment, "submit");
 
       // Provide a valid enrollment message.
       await this.core._handleMessage(
         {type: "enrollment", data: {}}
       );
 
-      assert.equal(await rallyMetrics.id.testGetValue(), FAKE_UUID);
+      assert.equal(await rallyMetrics.id.testGetValue("enrollment"), FAKE_UUID);
 
       // We expect to store the fake ion ID.
       assert.ok(this.core._storage.setRallyID.withArgs(FAKE_UUID).calledOnce);
       assert.ok(this.core._dataCollection.sendEnrollmentPing.calledOnce);
+      assert.ok(enrollmentPingSpy.calledOnce);
     });
 
     it('dispatches study-enrollment messages', async function () {
@@ -225,6 +233,8 @@ describe('Core', function () {
       chrome.runtime.getURL.returns(TEST_OPTIONS_URL);
 
       sinon.spy(this.core._dataCollection, "sendEnrollmentPing");
+      const enrollmentPingMock = sinon.mock(rallyPings.studyEnrollment);
+      enrollmentPingMock.expects("submit").once();
 
       // Mock the storage to provide a fake rally id.
       const FAKE_UUID = "c0ffeec0-ffee-c0ff-eec0-ffeec0ffeec0";
@@ -236,10 +246,13 @@ describe('Core', function () {
       // Attempt to enroll to a study.
       await this.core._enrollStudy(FAKE_STUDY_ID);
 
+      assert.equal(await enrollmentMetrics.studyId.testGetValue("study-enrollment"), FAKE_STUDY_ID);
+
       // We expect to store the fake ion ID.
       assert.ok(
-        this.core._dataCollection.sendEnrollmentPing.withArgs(FAKE_UUID, FAKE_STUDY_ID).calledOnce
+        this.core._dataCollection.sendEnrollmentPing.withArgs(FAKE_UUID, FAKE_STUDY_NAMESPACE).calledOnce
       );
+      enrollmentPingMock.verify();
     });
 
     it('dispatches unenrollment messages', async function () {
@@ -273,7 +286,7 @@ describe('Core', function () {
       assert.ok(this.core._storage.clearDeletionID.calledOnce);
       // ... to submit a ping with the expected type ...
       assert.ok(
-        this.core._dataCollection.sendDeletionPing.withArgs(FAKE_UUID, FAKE_STUDY_ID).calledOnce
+        this.core._dataCollection.sendDeletionPing.withArgs(FAKE_UUID, FAKE_STUDY_NAMESPACE).calledOnce
       );
       // We also expect an "uninstall" message to be dispatched to
       // the one study marked as installed.
@@ -324,7 +337,7 @@ describe('Core', function () {
 
       // We to submit a ping with the expected type.
       assert.ok(
-        this.core._dataCollection.sendDeletionPing.withArgs(FAKE_UUID, FAKE_STUDY_ID).calledOnce
+        this.core._dataCollection.sendDeletionPing.withArgs(FAKE_UUID, FAKE_STUDY_NAMESPACE).calledOnce
       );
 
       // Make sure that we're generating an uninstall message for
@@ -406,7 +419,7 @@ describe('Core', function () {
               FAKE_UUID,
               SENT_PING.payloadType,
               sinon.match(SENT_PING.payload),
-              SENT_PING.namespace,
+              FAKE_STUDY_NAMESPACE,
               SENT_PING.keyId,
               sinon.match(SENT_PING.key)
             ).calledOnce